perf(core): prioritize checking the producer that marked a consumer as dirty
Each consumer that has been marked as stale has to verify if one of its
producers has actually changed in the pull phase of the reactivity algorithm.
This involves iterating over all producers to see if any of them reports an
actual value change.

This commit implements an optimization where the producer that marks a consumer
as dirty will be captured in the consumer, such that the consumer can
prioritize checking that particular producer. This is beneficial, since it is
a likely candidate to actually report a change.