perf(core): store producers and consumers in a linear array structure
This commit swaps out the `Map` storage for producers and consumers to use a
linear array as data structure, with keys being stored at even indices and
values at odd indices. The primary benefit of this data structure is that it
provides faster iteration.

Deleting elements inserts `null` values into both the key and value slots,
meaning that the array never shrinks in size. These empty slots are reused
when inserting new elements.